requirements.txt
asgiref==3.7.2
autopep8==1.6.0
beautifulsoup4==4.12.2
bleach==6.0.0
crispy-bootstrap4==2022.1
Django==4.2.2
django-axes==6.3.0
django-bootstrap-v5==1.0.11
django-ckeditor==6.7.1
django-cleanup==6.0.0
django-crispy-forms==2.0
django-extensions==3.2.3
django-froala-editor==4.0.14
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-summernote==0.8.20.0
MarkupSafe==2.1.3
Pillow==10.0.0
pycodestyle==2.8.0
pydot==1.4.2
pyparsing==3.1.1
pytz==2024.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.6.3
tzdata==2022.1
webencodings==0.5.1
Werkzeug==2.3.7
